+++ src/index.ts
@@ -6,3 +6,4 @@ export * from "./bundle.js";
 export * from "./SessionRecorderLite.js";
 export * from "./OpenClawPassportWrapperLite.js";
 export * from "./OpenClawPassportMiddlewareLite.js";
+export * from "./html-export.js";